# Conflicts:
#	Game1007Project/Game1007Project/Game1007Project.vcxproj.filters
#	Game1007Project/Game1007Project/data/prefab/enemy2.xml
#	Game1007Project/Game1007Project/include/enemyShip.h

+++ Game1007Project/Game1007Project/source/script/projectile/straightProjectile.cpp
@@ -14,7 +14,7 @@ void StraightProjectile::Update()
 		GameObjectManager::GetInstance()->PopGameObject(gameObject);
 	}
 
-	int moveX, moveY;
+	float moveX, moveY;
 	GetMovePixel(&itemAttribute.vector, itemAttribute.speed, &moveX, &moveY);
 	this->transform->localPosition.y += moveY;
 	this->transform->localPosition.x += moveX;
